fix: localhost for development
