+++ web/lib/toc.ts
@@ -2,7 +2,31 @@ import { toc } from 'mdast-util-toc'
 import { remark } from 'remark'
 import { visit } from 'unist-util-visit'
 
-import { UnistNode } from './unist'
+type Node = import('unist').Node
+
+export interface UnistNode extends Node {
+  url?: string
+  type: string
+  name?: string
+  tagName?: string
+  value?: string
+  properties?: {
+    __rawString__?: string
+    __className__?: string
+    className?: string[]
+    [key: string]: unknown
+  }
+  attributes?: {
+    name: string
+    value: unknown
+    type?: string
+  }[]
+  children?: UnistNode[]
+}
+
+export interface UnistTree extends Node {
+  children: UnistNode[]
+}
 
 const textTypes = ['text', 'emphasis', 'strong', 'inlineCode']
