--- web/lib/unist.ts
@@ -1,25 +1,0 @@
-type Node = import('unist').Node
-
-export interface UnistNode extends Node {
-  url?: string
-  type: string
-  name?: string
-  tagName?: string
-  value?: string
-  properties?: {
-    __rawString__?: string
-    __className__?: string
-    className?: string[]
-    [key: string]: unknown
-  }
-  attributes?: {
-    name: string
-    value: unknown
-    type?: string
-  }[]
-  children?: UnistNode[]
-}
-
-export interface UnistTree extends Node {
-  children: UnistNode[]
-}
